Udemy-curso-API/mercado-livro/src/test/kotlin/com/mercadolivro/service/PurchaseServiceTest.kt
package com.mercadolivro.service

import com.mercadolivro.ModelBuilderHelper.buildCustomer
import com.mercadolivro.ModelBuilderHelper.buildPurchase
import com.mercadolivro.events.PurchaseEvent
import com.mercadolivro.model.BookModel
import com.mercadolivro.model.CustomerModel
import com.mercadolivro.model.PurchaseModel
import com.mercadolivro.repository.PurchaseRepository
import io.mockk.*
import io.mockk.impl.annotations.InjectMockKs
import io.mockk.impl.annotations.MockK
import io.mockk.junit5.MockKExtension
import org.junit.jupiter.api.Assertions.*
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.extension.ExtendWith
import org.springframework.context.ApplicationEventPublisher
import java.awt.print.Book
import java.math.BigDecimal
import java.util.UUID


@ExtendWith(MockKExtension::class)
class PurchaseServiceTest{
    @MockK
    private lateinit var purchaseRepository : PurchaseRepository
    @MockK
    private lateinit var applicationEventPublisher: ApplicationEventPublisher


    @InjectMockKs
    private lateinit var purchaseService: PurchaseService

    val purchaseEventSlot = slot<PurchaseEvent>()

    @Test
    fun `should create purchase and publish event`(){
        val purchase = buildPurchase()

        every {purchaseRepository.save(purchase) }returns purchase
        every { applicationEventPublisher.publishEvent(any()) } just runs

        purchaseService.create(purchase)

        verify(exactly = 1) {purchaseRepository.save(purchase)  }
        verify(exactly = 1) {applicationEventPublisher.publishEvent(capture(purchaseEventSlot))  }

        assertEquals(purchase, purchaseEventSlot.captured.purchaseModel)
    }

    @Test
    fun `should update purchase`(){
        val purchase = buildPurchase()

        every { purchaseRepository.save(purchase) }returns purchase

        purchaseService.update(purchase)

        verify(exactly = 1) {purchaseRepository.save(purchase)}
    }

}